basic tutorial moved to example folder
