+++ ode_robots/examples/basic/Makefile.conf
@@ -1,0 +1,9 @@
+# Configuration for simulation makefile
+# Please add all cpp files you want to compile for this simulation
+#  to the FILES variable
+# You can also tell where you haved lpzrobots installed
+
+FILES      = main differential basiccontroller
+LPZROBOTS  = ../../..
+ODEROBOTS  = $(LPZROBOTS)/ode_robots
+SELFORG    = $(LPZROBOTS)/selforg
